requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosmtplib==2.0.2
anyio==3.7.1
asgiref==3.6.0
asttokens==2.4.1
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
blis==0.7.10
blosc2==2.0.0
cachetools==5.3.0
catalogue==2.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
confection==0.1.1
cryptography==41.0.1
cymem==2.0.7
Cython==0.29.33
dataclasses-json==0.5.7
decorator==5.1.1
distlib==0.3.7
dnspython==2.3.0
docopt==0.6.2
ecdsa==0.18.0
email-validator==1.3.1
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.99.1
fastapi-jwt-auth==0.5.0
fastapi-mail==1.3.1
ffmpeg==1.4
filelock==3.12.2
Flask==2.2.3
Flask-SocketIO==5.3.3
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-language==2.10.1
google-cloud-speech==2.22.0
google-cloud-vision==3.4.4
googleapis-common-protos==1.58.0
gptcache==0.1.10
greenlet==2.0.2
grpcio==1.56.2
grpcio-status==1.56.2
gunicorn==20.1.0
h11==0.14.0
httplib2==0.21.0
httptools==0.5.0
idna==3.4
install==1.3.5
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
langchain==0.0.139
langcodes==3.3.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
msgpack==1.0.4
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.4
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.0
pandas==1.5.3
parso==0.8.3
passlib==1.7.4
pathy==0.10.2
pd==0.0.4
pipreqs==0.4.13
platformdirs==3.10.0
pods==0.1.14
preshed==3.0.8
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.22.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pydub==0.25.1
Pygments==2.16.1
PyJWT==1.7.1
pymongo==4.4.0
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-engineio==4.4.1
python-jose==3.3.0
python-multipart==0.0.5
python-socketio==5.8.0
pytz==2022.7.1
PyYAML==6.0
regex==2023.6.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.47
srsly==2.4.7
stack-data==0.6.3
starlette==0.27.0
tables==3.8.0
tenacity==8.2.2
thinc==8.1.12
tiktoken==0.4.0
tqdm==4.65.0
traitlets==5.13.0
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.22.0
virtualenv==20.24.2
wasabi==1.1.2
watchfiles==0.19.0
wcwidth==0.2.9
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
win32-setctime==1.1.0
xx-ent-wiki-sm @ https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-3.6.0/xx_ent_wiki_sm-3.6.0-py3-none-any.whl#sha256=5cd25405727477ac46ed96b511d8c893e0a734bb72e749967e2fa82db3005b4d
yarg==0.1.9
yarl==1.8.2
youtube-transcript-api==0.6.1
